chore: Bump Kubernetes (and associated) dependencies
This isn't going to pass. I'm sure of it. openapi gen is weirdly
messed up, I think `jx start pipeline` with Prow may not work, and
there's a good chance that Tekton won't work in general because of
bumping it to 0.9.2 - we may need to backport 0.9.2 syntax changes to
our 0.8.0 fork for the forklift up. But hey, this is an experiment and
I just wanted to see how it does.

This relates to #6957 kinda, and derives from the last round of
experimenting I did in #6413.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>